Moved Scanner functions and ScannerType into seperate file (mixin)
